+++ .github/workflows/deploy-sdk.yml
@@ -160,6 +160,8 @@ jobs:
         run: ./scripts/push-cocoapod.sh CustomerIOCommon.podspec
       - name: Push CustomerIOTracking
         run: ./scripts/push-cocoapod.sh CustomerIOTracking.podspec
+      - name: Push CustomerIODataPipelines
+        run: ./scripts/push-cocoapod.sh CustomerIODataPipelines.podspec
       - name: Push CustomerIOMessagingPush 
         run: ./scripts/push-cocoapod.sh CustomerIOMessagingPush.podspec
       - name: Push CustomerIOMessagingPushAPN
